# 플레이어의 액션(뛰기, 회피, 공격등)에 스테미나 제약 추가

+++ Assets/Scripts/UI/Player UI/StaminaBar.cs
@@ -9,13 +9,13 @@ public class StaminaBar : MonoBehaviour
     [SerializeField]
     private Slider slider;
 
-    public void SetMaxStamina(int maxStamina)
+    public void SetMaxStamina(float maxStamina)
     {
         slider.maxValue = maxStamina;
         slider.value = maxStamina;
     }
 
-    public void SetCurrentStamina(int curStamina)
+    public void SetCurrentStamina(float curStamina)
     {
         slider.value = curStamina;
     }
